Use title_with_badges on show pages and dashboard (and move long helpers out of application_helper)
