+++ app/views/dashboard/_community_news_item.html.erb
@@ -15,13 +15,13 @@
         <%= news.title %>
       </h3>
 
-      <p class="text-gray-600 text-sm mt-1 line-clamp-3">
+<!--      <p class="text-gray-600 text-sm mt-1 line-clamp-3">-->
         <%#= strip_tags(news.body) %>
-      </p>
+<!--      </p>-->
 
-      <p class="text-xs text-gray-400 mt-2">
+<!--      <p class="text-xs text-gray-400 mt-2">-->
         <%#= news.created_at.to_date.to_s(:long) %>
-      </p>
+<!--      </p>-->
     </div>
 
   <% end %>
