+++ app/views/quotes/index.html.erb
@@ -44,7 +44,7 @@
 <!--                      <span-->
 <!--                        class="inline-flex items-center px-2 py-0.5 rounded-full-->
 <!--                               text-xs font-medium bg-yellow-100 text-yellow-800 whitespace-nowrap">-->
-<!--                        🌟 Dashboard Feature -->
+<!--                        🌟 Dashboard feature -->
 <!--                      </span>-->
                     <%# end %>
                   </div>
