+++ app/views/resources/_resource_card.html.erb
@@ -44,9 +44,9 @@
     </div>
 
     <!-- RIGHT SIDE -->
-    <div class="flex-1 min-w-0 pl-4 pr-10 pb-12 relative">
+    <div class="flex-1 min-w-0 pl-4 pr-10 pb-12 mt-2 relative">
 
-      <div class="mb-2">
+      <div class="my-2">
         <%= link_to resource_path(resource.object),
                     class: "inline-flex min-w-0 max-w-full text-lg font-semibold
                             text-gray-900 leading-tight hover:underline" do %>
